fix(darkMode): fix dark mode issues
