+++ src/App.tsx
@@ -1,8 +1,13 @@
 import { RouterProvider } from 'react-router-dom';
 import Router from './router/Router';
+import { ThemeProvider } from './themes/theme.provider';
 
 const App = () => {
-  return <RouterProvider router={Router} />;
+  return (
+    <ThemeProvider defaultTheme="light" storageKey="vite-ui-theme">
+      <RouterProvider router={Router} />;
+    </ThemeProvider>
+  );
 };
 
 export default App;
