--- src/themes/theme-context.ts
@@ -1,13 +1,0 @@
-import { createContext } from 'react';
-
-export type Theme = 'dark' | 'light' | 'system';
-
-export type ThemeProviderState = {
-  theme: Theme;
-  setTheme: (theme: Theme) => void;
-};
-
-export const ThemeProviderContext = createContext<ThemeProviderState>({
-  theme: 'system',
-  setTheme: () => null,
-});
